# web-dev-projects/connectrix/connectrix-ui/src/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/messages":{"post":{"tags":["Message"],"operationId":"saveMessage","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageRequest"}}},"required":true},"responses":{"201":{"description":"Created"}}},"patch":{"tags":["Message"],"operationId":"setMessageToSeen","parameters":[{"name":"chat-id","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"202":{"description":"Accepted"}}}},"/api/v1/messages/upload-media":{"post":{"tags":["Message"],"operationId":"uploadMedia","parameters":[{"name":"chat-id","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":"Created"}}}},"/api/v1/chats":{"get":{"tags":["Chat"],"operationId":"getChatsByReceiver","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChatResponse"}}}}}}},"post":{"tags":["Chat"],"operationId":"createChat","parameters":[{"name":"sender-id","in":"query","required":true,"schema":{"type":"string"}},{"name":"receiver-id","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StringResponse"}}}}}}},"/api/v1/users":{"get":{"tags":["User"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}}}},"/api/v1/messages/chat/{chat-id}":{"get":{"tags":["Message"],"operationId":"getAllMessages","parameters":[{"name":"chat-id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageResponse"}}}}}}}}},"components":{"schemas":{"MessageRequest":{"type":"object","properties":{"content":{"type":"string"},"senderId":{"type":"string"},"receiverId":{"type":"string"},"type":{"type":"string","enum":["TEXT","IMAGE","VIDEO","AUDIO"]},"chatId":{"type":"string"}}},"StringResponse":{"type":"object","properties":{"response":{"type":"string"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"lastSeen":{"type":"string","format":"date-time"},"online":{"type":"boolean"}}},"MessageResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"content":{"type":"string"},"type":{"type":"string","enum":["TEXT","IMAGE","VIDEO","AUDIO"]},"state":{"type":"string","enum":["SENT","SEEN"]},"senderId":{"type":"string"},"receiverId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"media":{"type":"array","items":{"type":"string","format":"byte"}}}},"ChatResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"unreadCount":{"type":"integer","format":"int64"},"lastMessage":{"type":"string"},"lastMessageTime":{"type":"string","format":"date-time"},"senderId":{"type":"string"},"receiverId":{"type":"string"},"recipientOnline":{"type":"boolean"}}}},"securitySchemes":{"keycloak":{"type":"oauth2","in":"header","scheme":"bearer","bearerFormat":"JWT","flows":{"password":{"authorizationUrl":"http://localhost:9090/realms/connectrix/protocol/openid-connect/auth","tokenUrl":"http://localhost:9090/realms/connectrix/protocol/openid-connect/token","scopes":{}}}}}}}